Moved most ingredient search functionality to index.js Renamed fridgeDatabaeTest to ingredientApiSearch
